+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+## Unreleased
+* Add inline record support (issue #1)
+
 ## 0.1.1
 * Fix compatibility with OCaml 4.05, 4.06 and 4.07.